Bug fix from Eric Rogers:
"mod_fv3_lola.f90" initially aborted when the search algorithm to find the closest anl grid point to the fv3 grid point tried to reach outside the fv3 grid, aborting the run. The fix was just to put in a check to prevent this.

This makes NA 3km GSI analysis be able to use the 2x for analysis grid.